Update codeplug-format.md
Sorry for messing around with your documentation. For the first example, I've adopted my usual code-style for the code-plugs implementation. I just noticed that there is a new C++ type allowing to specify the underlying type and therefore size of an enum. This allows to specify fields in the code-plug with some restricted values. What do you think?